# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank Cards Management API
  description: REST API for managing bank cards, users, and transfers
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/cards:
    post:
      tags: [Cards]
      summary: Create a new card
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
      responses:
        '201':
          description: Card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
    get:
      tags: [Cards]
      summary: Get user cards (paginated)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Paginated cards
  /api/cards/{id}:
    get:
      tags: [Cards]
      summary: Get card by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Card found
    delete:
      tags: [Cards]
      summary: Delete card
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Card deleted
  /api/cards/{id}/block:
    patch:
      tags: [Cards]
      summary: Block card
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Card blocked
  /api/transfers:
    post:
      tags: [Transfers]
      summary: Transfer money between own cards
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Transfer completed
  /api/admin/users:
    get:
      tags: [Admin]
      summary: Get all users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
  /api/admin/users/{id}:
    get:
      tags: [Admin]
      summary: Get user by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User found
    delete:
      tags: [Admin]
      summary: Delete user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted
  /api/admin/users/{id}/role:
    patch:
      tags: [Admin]
      summary: Assign admin role
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Admin role assigned
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
    CardRequest:
      type: object
      properties:
        cardNumber:
          type: string
        cardHolder:
          type: string
        expiryDate:
          type: string
        initialBalance:
          type: number
    CardResponse:
      type: object
      properties:
        id:
          type: integer
        maskedNumber:
          type: string
        cardHolder:
          type: string
        expiryDate:
          type: string
        status:
          type: string
        balance:
          type: number
        ownerId:
          type: integer
    TransferRequest:
      type: object
      properties:
        fromCardId:
          type: integer
        toCardId:
          type: integer
        amount:
          type: number