# src/main/java/com/example/bankcards/dto/AuthResponse.java
package com.example.bankcards.dto;


public class AuthResponse {
    private String token;

    public AuthResponse(String token) {
        this.token = token;
    }

    public String getToken() {
        return token;
    }
}